+++ src/example/themes/LightTheme.ts
@@ -1,34 +1,8 @@
 import {Theme as NavigatorTheme} from '@react-navigation/native';
 import {StyleSheet} from 'react-native';
 
+import {ColorPalette} from './ColorPalette';
 import {Theme} from './react-native-elements';
-
-const ColorPalette = {
-  white: '#fff',
-  gray100: '#f8f9fa',
-  gray200: '#e9ecef',
-  gray300: '#dee2e6',
-  gray400: '#ced4da',
-  gray500: '#adb5bd',
-  gray600: '#6c757d',
-  gray700: '#495057',
-  gray800: '#343a40',
-  gray900: '#212529',
-  black: '#000',
-  blue: '#0d6efd',
-  indigo: '#6610f2',
-  purple: '#6f42c1',
-  pink: '#d63384',
-  red: '#dc3545',
-  orange: '#fd7e14',
-  yellow: '#ffc107',
-  green: '#198754',
-  teal: '#20c997',
-  cyan: '#0dcaf0',
-  get darkGray() {
-    return this.gray800;
-  },
-};
 
 const NavigatorColors: NavigatorTheme['colors'] = {
   primary: ColorPalette.blue,
@@ -44,7 +18,7 @@ export const LightTheme: Theme & NavigatorTheme = {
   colors: {
     ...NavigatorColors,
     white: '#fff',
-    black: '#000',
+    black: '#000', // 000はきつい 323232とかが
     secondary: '#ca71eb',
     grey0: '#393e42',
     grey1: '#43484d',
