+++ src/example/themes/SyncThemeWithDeviceSettings.tsx
@@ -1,25 +1,31 @@
 // テーマ切替用
 import {Theme as NavigatorTheme} from '@react-navigation/native';
-import React, {useEffect} from 'react';
-import {useColorScheme} from 'react-native';
+import React, {PropsWithChildren, useEffect} from 'react';
+import {ColorSchemeName, useColorScheme} from 'react-native';
 import {ThemeProps, withTheme} from 'react-native-elements';
 
 import {LightTheme, DarkTheme} from '.';
 import {useIsDark} from './IsDarkProvider';
 
-const SyncThemeWithDeviceSettingsInner: React.FC<ThemeProps<NavigatorTheme>> = (props) => {
+function isDark(scheme: ColorSchemeName) {
+  return scheme === 'dark';
+}
+
+function getTheme(scheme: ColorSchemeName) {
+  return isDark(scheme) ? DarkTheme : LightTheme;
+}
+
+function SyncThemeWithDeviceSettingsInner(props: PropsWithChildren<ThemeProps<NavigatorTheme>>) {
   const {updateTheme, children} = props;
   const {setIsDark} = useIsDark();
 
   const colorScheme = useColorScheme();
   useEffect(() => {
-    const nextIsDark = colorScheme?.toString() === 'dark';
-    const nextTheme = nextIsDark ? DarkTheme : LightTheme;
-    setIsDark(nextIsDark);
-    updateTheme(nextTheme);
+    setIsDark(isDark(colorScheme));
+    updateTheme(getTheme(colorScheme));
   }, [colorScheme]);
 
   return <>{children}</>;
-};
+}
 
 export const SyncThemeWithDeviceSettings = withTheme(SyncThemeWithDeviceSettingsInner);
